pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = ""
authors = ["Mohamed Benzerga <_mohamed.benzerga.az@sodexo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
langchain = "^0.1.16"
black = "^24.4.0"
openai = "^1.20.0"
install = "^1.3.5"
google-search-results = "^2.4.2"
tweepy = "^4.14.0"
flask = "^3.0.3"
langchain-openai = "^0.1.3"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.33"
pydantic = "^2.7.0"
langchain-core = "^0.1.43"
pylint = "^3.1.0"
langchainhub = "^0.1.15"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
